main.py
import discord
from discord.ext import commands
import os
import logging
import logging.handlers

from dotenv import load_dotenv
from schedule_cog import ScheduleCog 

load_dotenv()

logger = logging.getLogger('discord')
logger.setLevel(logging.DEBUG)
logging.getLogger('discord.http').setLevel(logging.INFO)

handler = logging.handlers.RotatingFileHandler(
    filename='discord.log',
    encoding='utf-8',
    maxBytes=32 * 1024 * 1024,
    backupCount=5,
)
dt_fmt = '%Y-%m-%d %H:%M:%S'
formatter = logging.Formatter('[{asctime}] [{levelname:<8}] {name}: {message}', dt_fmt, style='{')
handler.setFormatter(formatter)
logger.addHandler(handler)

TOKEN = os.getenv('DISCORD_BOT_TOKEN')


class BotClient(commands.Bot):
    def __init__(self):
        intents = discord.Intents.default()
        intents.message_content = True
        super().__init__(command_prefix='!', intents=intents)

    async def on_ready(self):
        print(f'Logged in as {self.user} (ID: {self.user.id})')
        
        await self.add_cog(ScheduleCog(self))
        
        try:
            synced = await self.tree.sync()
            print(f"Synced {len(synced)} slash command(s).")
        except Exception as e:
            print(f"Failed to sync commands: {e}")

        
if __name__ == '__main__':
    bot_client = BotClient()
    bot_client.run(TOKEN, log_handler=handler)
