pyproject.toml
[project]
name = "1879-schedule-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
]
